[ci] Run Jenkins agent with user 'jenkins' instead of root
Running the agent with root causes issues with our bind mount.
The repo ends up being owned by root, and as a result the agent cannot
run CI checks on the repository.

I tried quite a bit to switch the ownership via user namespace, but
unfortunately they did not work.
Therefore, I took the other way and checked what needs to be run as
root, and it was just SSHD since dind has a rootless variant which we
can use easily.

Therefore, the Containerfile and the entrypoint of our jenkins-agent is
updated to run SSHD as non-root user 'jenkins'.
Since the container runs as user 'jenkins', Podman automatically sets
the ownership of our bind mount (repo) to 'jenkins', which fixes the
issue we encountered before.