[ci] Create Makefile to connect pipeline with the project itself
A new Makefile is created to create an interface between the pipeline
and the actual project.
Seperating the actual CI commands from the Jenkinsfile allows us to keep
the pipeline simple, it essentially becomes a script runner and each
script (make targets) can be tested quite easily locally.

Also, the project's package naming is updated and the build image is
changed to golangci-lint, since it contains both the Go runtime and the
linter itself.