[ci] Create a custom Jenkins controller via casc
Up until this point, the testing of this project was done on a regular
Docker image.
The controller configurations were actually stored in a volume called
jenkins_controller.

In order to make a persistent configuration on the controller, a new
Containerfile is added to the project under
./.pipeline/local/jenkins-controller.

This Containerfile will contain the base Jenkins image and then the
actual configuration will be applied via Configuration as Code (CasC)
plugin.

So, instead of creating a backup of the current volume and applying it
to the container itself, a new custom container is created with
pre-defined configuration values via config.yml.

Currently, only users are created under config.yml. The rest of the
features (multibranch pipelines, repo setup, git plugin setup, etc.)
will be added incrementally.