[ci] Finalize the configuration of Jenkins controller via casc
Here is a list of changes that are done in this commit:

- The Containerfile of our controller is finalized: the steps are
ordered to take advantage of the layer caching process. Also, the
required plugin list is updated to fully customize our PoC environment
via configuration-as-code plugin.
- The configuration of the contoller is finalized: The current
configuration is written in a way that provides one permanent node, a
multibranch pipeline, and a global SSH credential that our controller
needs to connect to the node.
- A new file called local-multibranch.groovy is added to the project
because the only way to define jobs in YAML configuration is through
custom scripts as of now.

The build is tested via UI and the pipeline is executed without any
errors.